requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
psycopg2==2.9.9
pyasn1==0.5.1
pydantic==1.10.13
pydantic_core==2.16.2
pytest==8.0.1
pytest-asyncio==0.23.5
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
